requirements.txt
asgiref==3.2.10
Django==3.1.2
django-crispy-forms==1.10.0
django-filter==2.4.0
django-widget-tweaks==1.4.8
Pillow==8.0.1
pytz==2020.1
sqlparse==0.4.1
